+++ Bot/Cogs/mcsrvstats.py
@@ -1,6 +1,6 @@
 import aiohttp
 import discord
-import ujson
+import orjson
 from discord.ext import commands
 
 
@@ -10,7 +10,7 @@ class mcsrvstats(commands.Cog):
 
     @commands.command(name="javamcsrv", aliases=["java"])
     async def java(self, ctx, search: str):
-        async with aiohttp.ClientSession(json_serialize=ujson.dumps) as session:
+        async with aiohttp.ClientSession(json_serialize=orjson.dumps) as session:
             async with session.get(f"https://api.mcsrvstat.us/2/{search}") as r:
                 mcsrv = await r.json()
                 image_link = f"https://api.mcsrvstat.us/icon/{search}"
@@ -176,7 +176,7 @@ class bedrock_mcsrvstats(commands.Cog):
 
     @commands.command(name="bedrockmcsrv", aliases=["bedrock"])
     async def bedrock(self, ctx, search: str):
-        async with aiohttp.ClientSession(json_serialize=ujson.loads) as session:
+        async with aiohttp.ClientSession(json_serialize=orjson.loads) as session:
             async with session.get(f"https://api.mcsrvstat.us/bedrock/2/{search}") as r:
                 bedmcsrv = await r.json()
                 bedimage_link = f"https://api.mcsrvstat.us/icon/{search}"