+++ Bot/Cogs/waifu-pics.py
@@ -2,7 +2,7 @@ import random
 
 import aiohttp
 import discord
-import ujson
+import orjson
 from discord.ext import commands
 
 
@@ -46,11 +46,12 @@ class waifu(commands.Cog):
             "cringe",
         ]
         searchterm = random.choice(waifu_list)
-        async with aiohttp.ClientSession(json_serialize=ujson.dumps) as session:
+        async with aiohttp.ClientSession(json_serialize=orjson.dumps) as session:
             async with session.get(f"https://api.waifu.pics/sfw/{searchterm}") as r:
-                waifu_pics = await r.json()
+                waifu_pics = await r.text()
+                waifu_pics_formatted = orjson.loads(waifu_pics)
                 try:
-                    await ctx.send(waifu_pics["url"])
+                    await ctx.send(waifu_pics_formatted["url"])
                 except Exception as e:
                     embedVar = discord.Embed()
                     embedVar.description = "The query was not successful"